[UPDATE/1120] DustCloudGroup 기능 초안 업데이트

+++ KDH_DX2D_KZ/GameEngineContents/MainLevel2_3.cpp
@@ -258,7 +258,7 @@ void MainLevel2_3::LevelStart(GameEngineLevel* _PrevLevel)
 
 	{
 		std::shared_ptr<GameStateManager> Object = CreateActor<GameStateManager>();
-		Object->SetLeftEnemy(AllSpawnedEnemy.size());
+		Object->SetLeftEnemy(static_cast<int>(AllSpawnedEnemy.size()));
 	}
